fix: footer styles
